+++ webapp/src/app/app.module.ts
@@ -61,7 +61,11 @@ import { WebSocketService } from "../core/websocket.service";
 import { AvatarComponent } from "../components/avatar/avatar.component";
 import { Profile } from "../pages/profile/profile";
 import { AvatarService } from "../domain/avatar.service";
+import { TransactionQueueService } from "../domain/transaction-queue.service";
 import { ProgressBarComponent } from "../components/progress-bar/progress-bar.component";
+import { TwitterComponent } from "../components/twitter/twitter.component";
+import { FacebookComponent } from "../components/facebook/facebook.component";
+
 export function HttpLoaderFactory(http: HttpClient) {
     return new TranslateHttpLoader(http, "./assets/i18n/", ".json");
 }
@@ -90,7 +94,9 @@ export function HttpLoaderFactory(http: HttpClient) {
         FilterComponent,
         CustomRating,
         RankingCard,
-        ProgressBarComponent
+        ProgressBarComponent,
+        TwitterComponent,
+        FacebookComponent
     ],
     imports: [
         BrowserModule,
@@ -129,7 +135,9 @@ export function HttpLoaderFactory(http: HttpClient) {
         FilterComponent,
         CustomRating,
         RankingCard,
-        ProgressBarComponent
+        ProgressBarComponent,
+        TwitterComponent,
+        FacebookComponent
     ],
     providers: [
         AppConfig,
@@ -151,7 +159,8 @@ export function HttpLoaderFactory(http: HttpClient) {
         {provide: ErrorHandler, useClass: ErrorHandlerService},
         ContractManagerService,
         AppVersionService,
-        UpdateCheckService
+        UpdateCheckService,
+        TransactionQueueService
     ]
 })